gerenciamento_cliente/requirements.txt
asgiref==3.5.2
bottle==0.12.18
click==8.1.3
colorama==0.4.6
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.2
Werkzeug==2.2.3
